oox: avoid pointless multiplication at int precision in WpsContext
When the output would be used at long precision anyway.

Reviewed-on: https://gerrit.libreoffice.org/27360
Tested-by: Jenkins <ci@libreoffice.org>

+++ oox/source/shape/WpsContext.cxx
@@ -106,7 +106,7 @@ oox::core::ContextHandlerRef WpsContext::onCreateContext(sal_Int32 nElementToken
 
                     // If the text is not rotated the way the shape wants it already, set the angle.
                     const sal_Int32 nRotation = -270;
-                    if (static_cast<long>(basegfx::rad2deg(fRotate)) != NormAngle360(nRotation * 100) / 100)
+                    if (static_cast<long>(basegfx::rad2deg(fRotate)) != NormAngle360(static_cast<long>(nRotation) * 100) / 100)
                     {
                         comphelper::SequenceAsHashMap aCustomShapeGeometry(xPropertySet->getPropertyValue("CustomShapeGeometry"));
                         aCustomShapeGeometry["TextPreRotateAngle"] = uno::makeAny(nRotation);
